requirements.txt
# AutoID Validator - Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
playwright>=1.40.0
requests>=2.31.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.115.0

# Optional for better performance
# pywin32>=306; platform_system == "Windows"
